Create Day 9, Exercise 1 - Grading Program
100 Days of Code: The Complete Python Pro Bootcamp Coding Room Exercise.

Instructions:
You have access to a database of student_scores in the format of a dictionary. The keys in student_scores are the names of the students and the values are their exam scores. Write a program that converts their scores to grades. 

This is the scoring criteria:

Scores 91 - 100: Grade = "Outstanding"
Scores 81 - 90: Grade = "Exceeds Expectations"
Scores 71 - 80: Grade = "Acceptable"
Scores 70 or lower: Grade = "Fail"