Create Day 10, Exercise 1 - Days in Month
100 Days of Code: The Complete Python Pro Bootcamp Coding Room Exercise.

Instructions
In the starting code, you'll find the solution from the Leap Year challenge. First, convert this function is_leap() so that instead of printing "Leap year." or "Not leap year." it should return True if it is a leap year and return False if it is not a leap year.

You are then going to create a function called days_in_month() which will take a year and a month as inputs. 
And it will use this information to work out the number of days in the month, then return that as the output.